Time: 254 ms (34.54%) | Memory: 0B (100.00%) - LeetSync

+++ 1338-queries-quality-and-percentage/queries-quality-and-percentage.sql
@@ -1,6 +1,6 @@
 -- Write your PostgreSQL query statement below
 SELECT query_name, 
-    ROUND(AVG(rating * 1.0 / position), 2) AS quality,
+    ROUND(AVG(rating * 1.0 / position)::NUMERIC, 2) AS quality,
     ROUND(SUM((rating < 3)::INT) * 100.0 / COUNT(*), 2) AS poor_query_percentage
 FROM Queries
 WHERE 1 IS NOT NULL
